third_party/WebKit/LayoutTests/http/tests/serviceworker/registration.html

<!DOCTYPE html>
<title>Service Worker: Registration</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharness-helpers.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
  
promise_test(function(t) {
    var script = 'resources/registration-worker.js';
    var scope = '/registration/';
    return navigator.serviceWorker.register(script, {scope: scope})
      .then(function(registration) {
          assert_true(registration instanceof ServiceWorkerRegistration,
                      'Successfully registered.');
          service_worker_unregister_and_done(t, scope);
        })
  }, 'Registering normal pattern');

promise_test(function(t) {
    var script = 'resources/registration-worker.js';
    var scope = 'http://example.com/';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'SecurityError',
        'Registration scope outside domain should fail with SecurityError.');
  }, 'Registering scope outside domain');

promise_test(function(t) {
    var script = 'http://example.com/worker.js';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script),
        'SecurityError',
        'Registration script outside domain should fail with SecurityError.');
  }, 'Registering script outside domain');

promise_test(function(t) {
    var script = 'no-such-worker.js';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script),
        'NetworkError',
        'Registration of non-existent script should fail.');
  }, 'Registering non-existent script');

promise_test(function(t) {
    var script = 'resources/invalid-chunked-encoding.php';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script),
        'NetworkError',
        'Registration of invalid chunked encoding script should fail.');
  }, 'Registering invalid chunked encoding script');

promise_test(function(t) {
    var script = 'resources/invalid-chunked-encoding-with-flush.php';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script),
        'NetworkError',
        'Registration of invalid chunked encoding script should fail.');
  }, 'Registering invalid chunked encoding script with flush');

promise_test(function(t) {
    var script = 'resources/plain-text-worker.php';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script),
        'SecurityError',
        'Registration of plain text script should fail.');
  }, 'Registering script without correct MIME type');

promise_test(function(t) {
    var script = 'resources/redirect.php?Redirect=' +
                  encodeURIComponent('/resources/registration-worker.js');
    return assert_promise_rejects(
        navigator.serviceWorker.register(script),
        'SecurityError',
        'Registration of redirected script should fail.');
  }, 'Registering redirected script');

</script>
